make ukernel code inlinable down to arch-specific function pointer selection
