+++ runtime/src/iree/builtins/ukernel/arch/arm_64/BUILD.bazel
@@ -28,20 +28,40 @@ if(_IREE_UKERNEL_BITCODE_BUILD_ARM_64)
 # All headers transitively included by code in this directory. Bazel-only.
 UKERNEL_ARM_64_INTERNAL_HEADERS = [
     "common_arm_64.h",
+    "common_arm_64_entry_point.h",
+    "mmt4d_arm_64_internal.h",
+    "pack_arm_64_internal.h",
+    "unpack_arm_64_internal.h",
     "//runtime/src/iree/builtins/ukernel:internal_headers_filegroup",
     "//runtime/src/iree/schemas:cpu_data_headers_filegroup",
 ]
 
-UKERNEL_ARM_64_BASE_SRCS = [
-    "mmt4d_arm_64.c",
-    "pack_arm_64.c",
-    "query_tile_sizes_arm_64.c",
-    "unpack_arm_64.c",
-]
+iree_bitcode_library(
+    name = "ukernel_bitcode_arm_64_entry_points",
+    srcs = [
+        "mmt4d_arm_64_entry_point.c",
+        "pack_arm_64_entry_point.c",
+        "query_tile_sizes_arm_64_entry_point.c",
+        "unpack_arm_64_entry_point.c",
+    ],
+    # wasm_64 here is a proxy for "some reasonable 64-bit architecture". This
+    # should match the `ukernel_bitcode_64bit_base` bitcode library, because we
+    # want this code inlined into the ukernel entry points so that code path
+    # selection logic can evaluate at compile time and unused code paths can be
+    # DCE'd. The entry points don't need to be architecture-specialized, all
+    # they do is return function pointers to the actual architecture-specialized
+    # symbols defined in the other iree_bitcode_library's below.
+    arch = "wasm_64",
+    internal_hdrs = UKERNEL_ARM_64_INTERNAL_HEADERS,
+)
 
 iree_bitcode_library(
     name = "ukernel_bitcode_arm_64_base",
-    srcs = UKERNEL_ARM_64_BASE_SRCS,
+    srcs = [
+        "mmt4d_arm_64.c",
+        "pack_arm_64.c",
+        "unpack_arm_64.c",
+    ],
     arch = "arm_64",
     internal_hdrs = UKERNEL_ARM_64_INTERNAL_HEADERS,
 )
@@ -75,6 +95,7 @@ iree_cmake_extra_content(
     content = """
 elseif(IREE_BUILD_COMPILER AND IREE_TARGET_BACKEND_LLVM_CPU)
 iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_arm_64.bc")
+iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_arm_64_entry_points.bc")
 endif()  # _IREE_UKERNEL_BITCODE_BUILD_ARM_64
 """,
     inline = True,
