+++ runtime/src/iree/builtins/ukernel/arch/arm_64/CMakeLists.txt
@@ -15,13 +15,24 @@ if(_IREE_UKERNEL_BITCODE_BUILD_ARM_64)
 
 iree_bitcode_library(
   NAME
+    ukernel_bitcode_arm_64_entry_points
+  ARCH
+    wasm_64
+  SRCS
+    "mmt4d_arm_64_entry_point.c"
+    "pack_arm_64_entry_point.c"
+    "query_tile_sizes_arm_64_entry_point.c"
+    "unpack_arm_64_entry_point.c"
+)
+
+iree_bitcode_library(
+  NAME
     ukernel_bitcode_arm_64_base
   ARCH
     arm_64
   SRCS
     "mmt4d_arm_64.c"
     "pack_arm_64.c"
-    "query_tile_sizes_arm_64.c"
     "unpack_arm_64.c"
 )
 
@@ -59,6 +70,7 @@ iree_link_bitcode(
 
 elseif(IREE_BUILD_COMPILER AND IREE_TARGET_BACKEND_LLVM_CPU)
 iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_arm_64.bc")
+iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_arm_64_entry_points.bc")
 endif()  # _IREE_UKERNEL_BITCODE_BUILD_ARM_64
 
 ### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
@@ -125,9 +137,12 @@ iree_cc_library(
   NAME
     arm_64
   SRCS
+    "mmt4d_arm_64_entry_point.c"
     "mmt4d_arm_64.c"
+    "pack_arm_64_entry_point.c"
     "pack_arm_64.c"
-    "query_tile_sizes_arm_64.c"
+    "query_tile_sizes_arm_64_entry_point.c"
+    "unpack_arm_64_entry_point.c"
     "unpack_arm_64.c"
   DEPS
     ::common_arm_64
