+++ runtime/src/iree/builtins/ukernel/arch/arm_64/pack_arm_64.c
@@ -5,9 +5,9 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "iree/builtins/ukernel/arch/arm_64/common_arm_64.h"
-#include "iree/builtins/ukernel/pack_internal.h"
+#include "iree/builtins/ukernel/arch/arm_64/pack_arm_64_internal.h"
 
-static void iree_uk_pack_tile_8x1_x8_arm_64_direct(
+void iree_uk_pack_tile_8x1_x8_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -34,7 +34,7 @@ static void iree_uk_pack_tile_8x1_x8_arm_64_direct(
   }
 }
 
-static void iree_uk_pack_tile_8x4_x8_arm_64_direct(
+void iree_uk_pack_tile_8x4_x8_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -58,7 +58,7 @@ static void iree_uk_pack_tile_8x4_x8_arm_64_direct(
   }
 }
 
-static void iree_uk_pack_tile_8x1_x32_arm_64_direct(
+void iree_uk_pack_tile_8x1_x32_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -72,7 +72,7 @@ static void iree_uk_pack_tile_8x1_x32_arm_64_direct(
                                          4);
 }
 
-static void iree_uk_pack_tile_8x8_x8_arm_64_direct(
+void iree_uk_pack_tile_8x8_x8_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -90,7 +90,7 @@ static void iree_uk_pack_tile_8x8_x8_arm_64_direct(
   }
 }
 
-static void iree_uk_pack_tile_8x1_x32_arm_64_transpose(
+void iree_uk_pack_tile_8x1_x32_arm_64_transpose(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -108,7 +108,7 @@ static void iree_uk_pack_tile_8x1_x32_arm_64_transpose(
   }
 }
 
-static void iree_uk_pack_tile_8x1_x8_arm_64_transpose(
+void iree_uk_pack_tile_8x1_x8_arm_64_transpose(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -134,7 +134,7 @@ static void iree_uk_pack_tile_8x1_x8_arm_64_transpose(
   }
 }
 
-static void iree_uk_pack_tile_8x4_x8_arm_64_transpose(
+void iree_uk_pack_tile_8x4_x8_arm_64_transpose(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -161,7 +161,7 @@ static void iree_uk_pack_tile_8x4_x8_arm_64_transpose(
   }
 }
 
-static void iree_uk_pack_tile_8x8_x8_arm_64_transpose(
+void iree_uk_pack_tile_8x8_x8_arm_64_transpose(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -181,7 +181,7 @@ static void iree_uk_pack_tile_8x8_x8_arm_64_transpose(
   }
 }
 
-static void iree_uk_pack_tile_8x8_x32_arm_64_direct(
+void iree_uk_pack_tile_8x8_x32_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride1, iree_uk_index_t in_stride0,
@@ -198,31 +198,4 @@ static void iree_uk_pack_tile_8x8_x32_arm_64_direct(
     out_ptr += 4 * out_stride1;
     in_ptr += 32;
   }
-}
-
-iree_uk_pack_tile_func_t iree_uk_pack_select_tile_func_arch(
-    const iree_uk_pack_params_t* params) {
-  // At the moment, as sum-reductions are not yet part of pack ops,
-  // no arithmetic whatsoever is being done here, so only the element type
-  // size matters, not the type itself.
-  iree_uk_pack_type_t pack_type = iree_uk_pack_type(params->flags);
-  int esize = iree_uk_type_size(iree_uk_pack_out_type(pack_type));
-  bool transpose = params->flags & IREE_UK_FLAG_PACK_TRANSPOSE_INNER;
-  if (esize == 4 && params->out_size2 == 8 && params->out_size3 == 8) {
-    // Currently only used for accumulators, which are never transposed.
-    return transpose ? 0 : iree_uk_pack_tile_8x8_x32_arm_64_direct;
-  } else if (esize == 4 && params->out_size2 == 8 && params->out_size3 == 1) {
-    return transpose ? iree_uk_pack_tile_8x1_x32_arm_64_transpose
-                     : iree_uk_pack_tile_8x1_x32_arm_64_direct;
-  } else if (esize == 1 && params->out_size2 == 8 && params->out_size3 == 1) {
-    return transpose ? iree_uk_pack_tile_8x1_x8_arm_64_transpose
-                     : iree_uk_pack_tile_8x1_x8_arm_64_direct;
-  } else if (esize == 1 && params->out_size2 == 8 && params->out_size3 == 4) {
-    return transpose ? iree_uk_pack_tile_8x4_x8_arm_64_transpose
-                     : iree_uk_pack_tile_8x4_x8_arm_64_direct;
-  } else if (esize == 1 && params->out_size2 == 8 && params->out_size3 == 8) {
-    return transpose ? iree_uk_pack_tile_8x8_x8_arm_64_transpose
-                     : iree_uk_pack_tile_8x8_x8_arm_64_direct;
-  }
-  return 0;
 }
