+++ runtime/src/iree/builtins/ukernel/arch/arm_64/pack_arm_64_internal.h
@@ -1,0 +1,22 @@
+// Copyright 2023 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_BUILTINS_UKERNEL_ARCH_ARM_64_PACK_ARM_64_INTERNAL_H_
+#define IREE_BUILTINS_UKERNEL_ARCH_ARM_64_PACK_ARM_64_INTERNAL_H_
+
+#include "iree/builtins/ukernel/pack_internal.h"
+
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x1_x8_arm_64_direct)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x4_x8_arm_64_direct)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x1_x32_arm_64_direct)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x8_x8_arm_64_direct)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x1_x32_arm_64_transpose)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x1_x8_arm_64_transpose)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x4_x8_arm_64_transpose)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x8_x8_arm_64_transpose)
+IREE_UK_PACK_TILE_FUNC_DECL(iree_uk_pack_tile_8x8_x32_arm_64_direct)
+
+#endif  // foIREE_BUILTINS_UKERNEL_ARCH_ARM_64_PACK_ARM_64_INTERNAL_H_
