+++ runtime/src/iree/builtins/ukernel/arch/arm_64/unpack_arm_64.c
@@ -5,9 +5,9 @@
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
 #include "iree/builtins/ukernel/arch/arm_64/common_arm_64.h"
-#include "iree/builtins/ukernel/unpack_internal.h"
+#include "iree/builtins/ukernel/arch/arm_64/unpack_arm_64_internal.h"
 
-static void iree_uk_unpack_tile_8x8_x32_arm_64_direct(
+void iree_uk_unpack_tile_8x8_x32_arm_64_direct(
     void* IREE_UK_RESTRICT out_tile_ptr,
     const void* IREE_UK_RESTRICT in_tile_ptr, iree_uk_index_t outer_size1,
     iree_uk_index_t out_stride0, iree_uk_index_t in_stride1,
@@ -24,17 +24,4 @@ static void iree_uk_unpack_tile_8x8_x32_arm_64_direct(
     out_ptr += 32;
     in_ptr += 4 * in_stride1;
   }
-}
-
-iree_uk_unpack_tile_func_t iree_uk_unpack_select_tile_func_arch(
-    const iree_uk_unpack_params_t* params) {
-  iree_uk_unpack_type_t unpack_type = iree_uk_unpack_type(params->flags);
-  int esize = iree_uk_type_size(iree_uk_unpack_out_type(unpack_type));
-  bool transpose = params->flags & IREE_UK_FLAG_UNPACK_TRANSPOSE_INNER;
-  // Unpack is currently only used in practice with esize==4 and non-transpose.
-  if (esize != 4 || transpose) return 0;
-  if (params->in_size2 == 8 && params->in_size3 == 8) {
-    return iree_uk_unpack_tile_8x8_x32_arm_64_direct;
-  }
-  return 0;
 }
