+++ runtime/src/iree/builtins/ukernel/arch/arm_64/unpack_arm_64_internal.h
@@ -1,0 +1,14 @@
+// Copyright 2023 The IREE Authors
+//
+// Licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#ifndef IREE_BUILTINS_UKERNEL_ARCH_ARM_64_UNPACK_ARM_64_INTERNAL_H_
+#define IREE_BUILTINS_UKERNEL_ARCH_ARM_64_UNPACK_ARM_64_INTERNAL_H_
+
+#include "iree/builtins/ukernel/unpack_internal.h"
+
+IREE_UK_UNPACK_TILE_FUNC_DECL(iree_uk_unpack_tile_8x8_x32_arm_64_direct)
+
+#endif  // IREE_BUILTINS_UKERNEL_ARCH_ARM_64_UNPACK_ARM_64_INTERNAL_H_
