+++ runtime/src/iree/builtins/ukernel/arch/x86_64/BUILD.bazel
@@ -28,29 +28,32 @@ if(_IREE_UKERNEL_BITCODE_BUILD_X86_64)
 # All headers transitively included by code in this directory. Bazel-only.
 UKERNEL_X86_64_INTERNAL_HEADERS = [
     "common_x86_64.h",
+    "common_x86_64_entry_point.h",
+    "mmt4d_x86_64_internal.h",
+    "pack_x86_64_internal.h",
+    "unpack_x86_64_internal.h",
     "//runtime/src/iree/builtins/ukernel:internal_headers_filegroup",
     "//runtime/src/iree/schemas:cpu_data_headers_filegroup",
 ]
 
-UKERNEL_X86_64_BASE_SRCS = [
-    "mmt4d_x86_64.c",
-    "pack_x86_64.c",
-    "query_tile_sizes_x86_64.c",
-    "unpack_x86_64.c",
-]
-
 iree_bitcode_library(
-    name = "ukernel_bitcode_x86_64_base",
-    srcs = UKERNEL_X86_64_BASE_SRCS,
-    arch = "x86_64",
+    name = "ukernel_bitcode_x86_64_entry_points",
+    srcs = [
+        "mmt4d_x86_64_entry_point.c",
+        "pack_x86_64_entry_point.c",
+        "query_tile_sizes_x86_64_entry_point.c",
+        "unpack_x86_64_entry_point.c",
+    ],
+    # wasm_64 here is a proxy for "some reasonable 64-bit architecture". This
+    # should match the `ukernel_bitcode_64bit_base` bitcode library, because we
+    # want this code inlined into the ukernel entry points so that code path
+    # selection logic can evaluate at compile time and unused code paths can be
+    # DCE'd. The entry points don't need to be architecture-specialized, all
+    # they do is return function pointers to the actual architecture-specialized
+    # symbols defined in the other iree_bitcode_library's below.
+    arch = "wasm_64",
     internal_hdrs = UKERNEL_X86_64_INTERNAL_HEADERS,
 )
-
-UKERNEL_X86_64_AVX2_FMA_SRCS = [
-    "mmt4d_x86_64_avx2_fma.c",
-    "pack_x86_64_avx2_fma.c",
-    "unpack_x86_64_avx2_fma.c",
-]
 
 UKERNEL_X86_64_AVX2_FMA_COPTS = [
     "-mavx",
@@ -60,17 +63,15 @@ UKERNEL_X86_64_AVX2_FMA_COPTS = [
 
 iree_bitcode_library(
     name = "ukernel_bitcode_x86_64_avx2_fma",
-    srcs = UKERNEL_X86_64_AVX2_FMA_SRCS,
+    srcs = [
+        "mmt4d_x86_64_avx2_fma.c",
+        "pack_x86_64_avx2_fma.c",
+        "unpack_x86_64_avx2_fma.c",
+    ],
     arch = "x86_64",
     copts = UKERNEL_X86_64_AVX2_FMA_COPTS,
     internal_hdrs = UKERNEL_X86_64_INTERNAL_HEADERS,
 )
-
-UKERNEL_X86_64_AVX512_BASE_SRCS = [
-    "mmt4d_x86_64_avx512_base.c",
-    "pack_x86_64_avx512_base.c",
-    "unpack_x86_64_avx512_base.c",
-]
 
 UKERNEL_X86_64_AVX512_BASE_COPTS = UKERNEL_X86_64_AVX2_FMA_COPTS + [
     "-mavx512f",
@@ -82,15 +83,15 @@ UKERNEL_X86_64_AVX512_BASE_COPTS = UKERNEL_X86_64_AVX2_FMA_COPTS + [
 
 iree_bitcode_library(
     name = "ukernel_bitcode_x86_64_avx512_base",
-    srcs = UKERNEL_X86_64_AVX512_BASE_SRCS,
+    srcs = [
+        "mmt4d_x86_64_avx512_base.c",
+        "pack_x86_64_avx512_base.c",
+        "unpack_x86_64_avx512_base.c",
+    ],
     arch = "x86_64",
     copts = UKERNEL_X86_64_AVX512_BASE_COPTS,
     internal_hdrs = UKERNEL_X86_64_INTERNAL_HEADERS,
 )
-
-UKERNEL_X86_64_AVX512_VNNI_SRCS = [
-    "mmt4d_x86_64_avx512_vnni.c",
-]
 
 UKERNEL_X86_64_AVX512_VNNI_COPTS = UKERNEL_X86_64_AVX512_BASE_COPTS + [
     "-mavx512vnni",
@@ -98,7 +99,9 @@ UKERNEL_X86_64_AVX512_VNNI_COPTS = UKERNEL_X86_64_AVX512_BASE_COPTS + [
 
 iree_bitcode_library(
     name = "ukernel_bitcode_x86_64_avx512_vnni",
-    srcs = UKERNEL_X86_64_AVX512_VNNI_SRCS,
+    srcs = [
+        "mmt4d_x86_64_avx512_vnni.c",
+    ],
     arch = "x86_64",
     copts = UKERNEL_X86_64_AVX512_VNNI_COPTS,
     internal_hdrs = UKERNEL_X86_64_INTERNAL_HEADERS,
@@ -107,7 +110,6 @@ iree_bitcode_library(
 iree_link_bitcode(
     name = "ukernel_bitcode_x86_64",
     bitcode_files = [
-        "ukernel_bitcode_x86_64_base.bc",
         "ukernel_bitcode_x86_64_avx2_fma.bc",
         "ukernel_bitcode_x86_64_avx512_base.bc",
         "ukernel_bitcode_x86_64_avx512_vnni.bc",
@@ -118,6 +120,7 @@ iree_cmake_extra_content(
     content = """
 elseif(IREE_BUILD_COMPILER AND IREE_TARGET_BACKEND_LLVM_CPU)
 iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_x86_64.bc")
+iree_make_empty_file("${CMAKE_CURRENT_BINARY_DIR}/ukernel_bitcode_x86_64_entry_points.bc")
 endif()  # _IREE_UKERNEL_BITCODE_BUILD_X86_64
 """,
     inline = True,
