+++ runtime/src/iree/builtins/ukernel/arch/x86_64/mmt4d_x86_64_entry_point.c
@@ -4,25 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/builtins/ukernel/arch/x86_64/common_x86_64.h"
-#include "iree/builtins/ukernel/mmt4d_internal.h"
-
-#if defined(IREE_UK_BUILD_X86_64_AVX2_FMA)
-IREE_UK_MMT4D_TILE_FUNC_DECL(iree_uk_mmt4d_tile_i8i8i32_8x8x2_x86_64_avx2_fma)
-IREE_UK_MMT4D_TILE_FUNC_DECL(iree_uk_mmt4d_tile_f32f32f32_8x8x1_x86_64_avx2_fma)
-#endif  // defined (IREE_UK_BUILD_X86_64_AVX2_FMA)
-
-#if defined(IREE_UK_BUILD_X86_64_AVX512_BASE)
-IREE_UK_MMT4D_TILE_FUNC_DECL(
-    iree_uk_mmt4d_tile_i8i8i32_16x16x2_x86_64_avx512_base)
-IREE_UK_MMT4D_TILE_FUNC_DECL(
-    iree_uk_mmt4d_tile_f32f32f32_16x16x1_x86_64_avx512_base)
-#endif  // defined (IREE_UK_BUILD_X86_64_AVX512_BASE)
-
-#if defined(IREE_UK_BUILD_X86_64_AVX512_VNNI)
-IREE_UK_MMT4D_TILE_FUNC_DECL(
-    iree_uk_mmt4d_tile_i8i8i32_16x16x2_x86_64_avx512_vnni)
-#endif  // defined (IREE_UK_BUILD_X86_64_AVX512_VNNI)
+#include "iree/builtins/ukernel/arch/x86_64/common_x86_64_entry_point.h"
+#include "iree/builtins/ukernel/arch/x86_64/mmt4d_x86_64_internal.h"
 
 static iree_uk_mmt4d_tile_func_t
 iree_uk_mmt4d_select_tile_func_x86_64_f32f32f32_8x8x1(
