+++ runtime/src/iree/builtins/ukernel/arch/x86_64/unpack_x86_64_entry_point.c
@@ -4,8 +4,8 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/builtins/ukernel/arch/x86_64/common_x86_64.h"
-#include "iree/builtins/ukernel/unpack_internal.h"
+#include "iree/builtins/ukernel/arch/x86_64/common_x86_64_entry_point.h"
+#include "iree/builtins/ukernel/arch/x86_64/unpack_x86_64_internal.h"
 
 IREE_UK_UNPACK_TILE_FUNC_DECL(
     iree_uk_unpack_tile_8x8_x32_x86_64_avx2_fma_direct)
